Update Virtualization (Again)
Apparently there was no problem with virtualization, rather, the problem
was updating the list so often.  LINQ queries were made on the addition
of every song (that's a problem, lol).  Added an event/delegate to the
API class, and upon paging the view is updated.  This is much faster,
but still not optimal as it's a little glitchy.